requirements.txt
tensorflow-gpu==1.15
nltk
pandas
tabulate
